+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 #### Changed
 * `html` - `SimpleTagHandler` visits children tags if supplied tag is block one ([#235])
 * `inline-parser` -  `BangInlineProcessor` properly returns `null` if no image node is found (possible to define other inline parsers that use `!` as special character)
+* `image` - `AsyncDrawable` won't trigger loading if it has result (aim: `RecyclerView` due to multiple attach/detach events of a View)
+* `image` - `AsyncDrawable` will resume result if it is `Animatable` and was playing before detach event (aim: `RecyclerView`) ([#241])
 
 #### Fixed
 * `image-glide` cache `RequestManager` in  `GlideImagesPlugin#create(Context)` factory method ([#259])
@@ -20,6 +22,7 @@
 * `image` - `AsyncDrawable#hasKnownDimentions` (deprecated in `4.2.1`)
 
 [#235]: https://github.com/noties/Markwon/issues/235
+[#241]: https://github.com/noties/Markwon/issues/241
 [#259]: https://github.com/noties/Markwon/issues/259